refactor: extract shared ScannerFactory to internal/scanner package
Create a shared scanner package that can be used by both:
- CDC query layer (internal/cdc)
- SQL plugin MSSQL executor (plugin/sql)

This ensures MSSQL DECIMAL/NUMERIC types are correctly converted
to strings using NumericString scanner, preventing BLOB/base64 issues.

Changes:
- internal/scanner: New package with ScannerFactory and all scanner types
- internal/cdc: Refactored to use scanner package, re-exports for compatibility
- plugin/sql/driver.go: MSSQLExecutor now uses scanner.ScannerFactory
